+++ src/Codeception/Lib/Connector/Phalcon.php
@@ -1,21 +1,18 @@
 <?php
 namespace Codeception\Lib\Connector;
 
-use Codeception\Lib\Connector\Shared\PhpSuperGlobalsConverter;
-use Symfony\Component\BrowserKit\Response;
+use Closure;
+use Phalcon\Di;
+use Phalcon\Http;
+use RuntimeException;
+use ReflectionProperty;
+use Codeception\Util\Stub;
+use Phalcon\Mvc\Application;
 use Symfony\Component\BrowserKit\Cookie;
 use Symfony\Component\BrowserKit\Client;
-use Codeception\Util\Stub;
-use Phalcon\Di;
-use Phalcon\Mvc\Application;
 use Phalcon\Mvc\Micro as MicroApplication;
-use Phalcon\Http\Request;
-use Phalcon\Http\RequestInterface;
-use Phalcon\Http\ResponseInterface;
-use Phalcon\Session\AdapterInterface as SessionInterface;
-use ReflectionProperty;
-use RuntimeException;
-use Closure;
+use Symfony\Component\BrowserKit\Response;
+use Codeception\Lib\Connector\Shared\PhpSuperGlobalsConverter;
 
 class Phalcon extends Client
 {
@@ -49,6 +46,7 @@ class Phalcon extends Client
         if ($application instanceof Closure) {
             return $application();
         } elseif (is_string($application)) {
+            /** @noinspection PhpIncludeInspection */
             return require $application;
         } else {
             return $application;
@@ -71,20 +69,20 @@ class Phalcon extends Client
         }
 
         $di = $application->getDI();
-        /** @var \Phalcon\Http\Request $phRequest */
+        /** @var Http\Request $phRequest */
         if ($di->has('request')) {
             $phRequest = $di->get('request');
         }
 
-        if (!$phRequest instanceof RequestInterface) {
-            $phRequest = new Request;
+        if (!$phRequest instanceof Http\RequestInterface) {
+            $phRequest = new Http\Request();
         }
 
         $uri         = $request->getUri() ?: $phRequest->getURI();
         $pathString  = parse_url($uri, PHP_URL_PATH);
         $queryString = parse_url($uri, PHP_URL_QUERY);
 
-        $_SERVER = $request->getServer();
+        $_SERVER = array_merge($_SERVER, $request->getServer());
         $_SERVER['REQUEST_METHOD'] = strtoupper($request->getMethod());
         $_SERVER['REQUEST_URI'] = null === $queryString ? $pathString : $pathString . '?' . $queryString;
 
@@ -114,7 +112,7 @@ class Phalcon extends Client
         $di['request'] = Stub::construct($phRequest, [], ['getRawBody' => $request->getContent()]);
 
         $response = $application->handle();
-        if (!$response instanceof ResponseInterface) {
+        if (!$response instanceof Http\ResponseInterface) {
             $response = $application->response;
         }
 
@@ -157,308 +155,5 @@ class Phalcon extends Client
             $status ? $status : 200,
             $headers
         );
-    }
-}
-
-class PhalconMemorySession implements SessionInterface
-{
-    /**
-     * @var string
-     */
-    protected $sessionId;
-
-    /**
-     * @var string
-     */
-    protected $name;
-
-    /**
-     * @var bool
-     */
-    protected $started = false;
-
-    /**
-     * @var array
-     */
-    protected $memory = [];
-
-    /**
-     * @var array
-     */
-    protected $options = [];
-
-    public function __construct(array $options = null)
-    {
-        $this->sessionId = $this->generateId();
-
-        if (is_array($options)) {
-            $this->setOptions($options);
-        }
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function start()
-    {
-        if ($this->status() !== PHP_SESSION_ACTIVE) {
-            $this->memory = [];
-            $this->started = true;
-
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param array $options
-     */
-    public function setOptions(array $options)
-    {
-        if (isset($options['uniqueId'])) {
-            $this->sessionId = $options['uniqueId'];
-        }
-
-        $this->options = $options;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @return array
-     */
-    public function getOptions()
-    {
-        return $this->options;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param string $index
-     * @param mixed $defaultValue
-     * @param bool $remove
-     * @return mixed
-     */
-    public function get($index, $defaultValue = null, $remove = false)
-    {
-        $key = $this->prepareIndex($index);
-
-        if (!isset($this->memory[$key])) {
-            return $defaultValue;
-        }
-
-        $return = $this->memory[$key];
-
-        if ($remove) {
-            unset($this->memory[$key]);
-        }
-
-        return $return;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param string $index
-     * @param mixed $value
-     */
-    public function set($index, $value)
-    {
-        $this->memory[$this->prepareIndex($index)] = $value;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param string $index
-     * @return bool
-     */
-    public function has($index)
-    {
-        return isset($this->memory[$this->prepareIndex($index)]);
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param string $index
-     */
-    public function remove($index)
-    {
-        unset($this->memory[$this->prepareIndex($index)]);
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @return string
-     */
-    public function getId()
-    {
-        return $this->sessionId;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @return bool
-     */
-    public function isStarted()
-    {
-        return $this->started;
-    }
-
-    /**
-     * Returns the status of the current session
-     *
-     * ``` php
-     * <?php
-     * if ($session->status() !== PHP_SESSION_ACTIVE) {
-     *     $session->start();
-     * }
-     * ?>
-     * ```
-     *
-     * @return int
-     */
-    public function status()
-    {
-        if ($this->isStarted()) {
-            return PHP_SESSION_ACTIVE;
-        }
-
-        return PHP_SESSION_NONE;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param bool $removeData
-     * @return bool
-     */
-    public function destroy($removeData = false)
-    {
-        if ($removeData) {
-            if (!empty($this->sessionId)) {
-                foreach ($this->memory as $key => $value) {
-                    if (0 === strpos($key, $this->sessionId . '#')) {
-                        unset($this->memory[$key]);
-                    }
-                }
-            } else {
-                $this->memory = [];
-            }
-        }
-
-        $this->started = false;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param bool $deleteOldSession
-     * @return \Phalcon\Session\AdapterInterface
-     */
-    public function regenerateId($deleteOldSession = true)
-    {
-        $this->sessionId = $this->generateId();
-
-        return $this;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param string $name
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @return string
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * Dump all session
-     *
-     * @return array
-     */
-    public function toArray()
-    {
-        return (array) $this->memory;
-    }
-
-    /**
-     * Alias: Gets a session variable from an application context
-     *
-     * @param string $index
-     * @return mixed
-     */
-    public function __get($index)
-    {
-        return $this->get($index);
-    }
-
-    /**
-     * Alias: Sets a session variable in an application context
-     *
-     * @param string $index
-     * @param mixed $value
-     */
-    public function __set($index, $value)
-    {
-        $this->set($index, $value);
-    }
-
-    /**
-     * Alias: Check whether a session variable is set in an application context
-     *
-     * @param  string $index
-     * @return bool
-     */
-    public function __isset($index)
-    {
-        return $this->has($index);
-    }
-
-    /**
-     * Alias: Removes a session variable from an application context
-     *
-     * @param string $index
-     */
-    public function __unset($index)
-    {
-        $this->remove($index);
-    }
-
-    private function prepareIndex($index)
-    {
-        if ($this->sessionId) {
-            $key = $this->sessionId . '#' . $index;
-        } else {
-            $key = $index;
-        }
-
-        return $key;
-    }
-
-    /**
-     * @return string
-     */
-    private function generateId()
-    {
-        return md5(time());
     }
 }