+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+#### 2.2.7
+
+* [Phalcon] Added `services` part which can be used to `grabServiceFromContainer` and `addServiceToContainer` when conflicting module is used. By @sergeyklay
+* [Phalcon] **Refactored**. Moved in-memory session adapter to the separated namespace. By @sergeyklay
+* [Phalcon] Fixed overwriting server parameters on requests. By @sergeyklay
+
 #### 2.2.6 (October 2016)
 
 * Ability to update config on run with `--override` (`-o`) option. Usage Examples: